requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr>=2.0.0 # Apache-2.0
Babel>=2.3.4 # BSD
Flask!=0.11,<1.0,>=0.10 # BSD
iso8601>=0.1.11 # MIT
jsonschema>=2.0.0,<3.0.0,!=2.5.0 # MIT
oslo.config>=3.22.0 # Apache-2.0
oslo.i18n>=2.1.0 # Apache-2.0
oslo.log>=3.22.0 # Apache-2.0
oslo.serialization>=1.10.0 # Apache-2.0
oslo.utils>=3.20.0 # Apache-2.0
paramiko>=2.0 # LGPLv2.1+
psutil>=3.2.2 # BSD
PyGithub
python-memcached>=1.56 # PSF
PyYAML>=3.10.0 # MIT
requests!=2.12.2,!=2.13.0,>=2.10.0 # Apache-2.0
requests-file
sh
six>=1.9.0 # MIT

# Mirantis
pillow
